Store coordinator instead of API in runtime_data, move API init to coordinator and remove storing/unloading to/from hass.data
